requirements.txt
alembic==0.9.6
aniso8601==8.0.0
azure-core==1.7.0
azure-storage-file-share==12.1.2
bcrypt==3.1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==6.7
coverage==4.4.2
cryptography==2.9.2
enum-compat==0.0.3
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-PyMongo==2.3.0
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
greenlet==0.4.16
gunicorn==19.7.1
idna==2.10
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
msrest==0.6.17
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
pymongo==3.10.1
python-dateutil==2.6.1
python-editor==1.0.4
pytz==2017.3
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==0.15.3
alembic==0.9.6
aniso8601==8.0.0
azure-core==1.7.0
azure-storage-file-share==12.1.2
bcrypt==3.1.4
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==6.7
coverage==4.4.2
cryptography==2.9.2
enum-compat==0.0.3
eventlet==0.21.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-PyMongo==2.3.0
flask-restplus==0.10.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
greenlet==0.4.16
gunicorn==19.7.1
idna==2.10
isodate==0.6.0
itsdangerous==0.24
Jinja2==2.11.2
jsonschema==2.6.0
Mako==1.0.7
MarkupSafe==1.1.1
msrest==0.6.17
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycparser==2.18
PyJWT==1.5.3
pymongo==3.10.1
python-dateutil==2.6.1
python-editor==1.0.4
pytz==2017.3
requests==2.24.0
requests-oauthlib==1.3.0
selenium==3.8.1
six==1.11.0
SQLAlchemy==1.3.18
urllib3==1.25.9
Werkzeug==0.15.3
xlrd==1.2.0
flask-restx==0.1.1
numpy==1.18.1
pandas==1.0.1

